0097 fix tests (#59)
* add all tests to PrimeTime app scheme

This change lets you run all tests (for all modules) using Product > Test with the PrimeTime scheme selected.

* fix tests for ViewStore refactoring

This commit updates CounterTests and PrimeTimeTests to work with CounterFeatureState and the corresponding changes to actions.

* update CounterTests snapshots

This commit updates the snapshot taken while the nth prime request is in
flight.

Originally, the + and - buttons were still enabled while the request is
in flight. Then the app was changed to disable them while the request is
in flight, which changes their appearance. This snapshot shows the
buttons with the disabled appearance.

Co-authored-by: Rob Mayoff <rmayoff@drwholdings.com>